backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
openai==1.57.2
python-multipart==0.0.20
python-dotenv==1.0.1
requests==2.32.3
